+++ build_modules/lib/build_modules.dart
@@ -5,6 +5,8 @@
 export 'src/errors.dart' show MissingModulesException;
 export 'src/kernel_builder.dart'
     show KernelSummaryBuilder, kernelSummaryExtension, multiRootScheme;
+export 'src/meta_module_builder.dart'
+    show MetaModuleBuilder, metaModuleExtension;
 export 'src/module_builder.dart' show ModuleBuilder, moduleExtension;
 export 'src/modules.dart';
 export 'src/scratch_space.dart' show scratchSpace, scratchSpaceResource;