+++ build_modules/lib/builders.dart
@@ -4,7 +4,10 @@
 
 import 'package:build/build.dart';
 import 'package:build_modules/build_modules.dart';
+import 'package:build_modules/src/meta_module_builder.dart';
 
-Builder moduleBuilder(_) => const ModuleBuilder();
+Builder moduleBuilder(BuilderOptions options) =>
+    new ModuleBuilder.forOptions(options);
 Builder unlinkedSummaryBuilder(_) => const UnlinkedSummaryBuilder();
 Builder linkedSummaryBuilder(_) => const LinkedSummaryBuilder();
+Builder metaModuleBuilder(_) => const MetaModuleBuilder();