+++ build_modules/lib/src/module_builder.dart
@@ -7,15 +7,119 @@ import 'dart:convert';
 
 import 'package:build/build.dart';
 
+import 'meta_module.dart';
+import 'meta_module_builder.dart';
 import 'modules.dart';
 
 /// The extension for serialized module assets.
 const moduleExtension = '.module';
 
+/// A map of primary [AssetId] to corresponding [Module].
+final _primaryToClean =
+    new Resource<Map<AssetId, Module>>(() => {}, dispose: (map) => map.clear());
+
+/// Set of all [MetaModule]s which have been processed.
+final _readMetas = new Resource<Set<AssetId>>(() => new Set<AssetId>(),
+    dispose: (s) => s.clear());
+
+/// Map of [AssetId] to corresponding primary [AssetId] within the same
+/// [Module].
+final _assetToPrimary = new Resource<Map<AssetId, AssetId>>(() => {},
+    dispose: (map) => map.clear());
+
+void _cacheAssetToPrimary(
+    MetaModule meta, Map<AssetId, AssetId> assetToPrimary) {
+  for (var module in meta.modules) {
+    for (var source in module.sources) {
+      assetToPrimary[source] = module.primarySource;
+    }
+  }
+}
+
+/// Updates dependencies from the provided [Module] so that they point to the
+/// primary resource of the dependencies corresponding [Module].
+Future<Module> _cleanModuleDeps(
+    BuildStep buildStep, Module module, Map<AssetId, AssetId> assetToPrimary,
+    {bool isCourseStrategy: false}) async {
+  var cleanedDeps = new Set<AssetId>();
+  for (var dep in module.directDependencies) {
+    if (!isCourseStrategy) {
+      // Since we are not using the ccourse strategy we can safely add
+      // all dependencies in the same package as they will have a
+      // corresponding module file.
+      if (dep.package == module.primarySource.package) {
+        cleanedDeps.add(dep);
+        continue;
+        // It is possible that this dep came from another fine grained
+        // module. Look for the corresponding module file.
+      } else if (await buildStep
+          .canRead(dep.changeExtension(moduleExtension))) {
+        cleanedDeps.add(dep);
+        continue;
+      }
+    }
+    var primaryDep = assetToPrimary[dep];
+    if (primaryDep == null) {
+      // We don't know the primary source for the dep so look for it
+      // using the meta module file.
+      var depAsset = new AssetId(dep.package, 'lib/$metaModuleExtension');
+      if (await buildStep.canRead(depAsset)) {
+        var depMeta = new MetaModule.fromJson(
+            json.decode(await buildStep.readAsString(depAsset))
+                as Map<String, dynamic>);
+        _cacheAssetToPrimary(depMeta, assetToPrimary);
+      }
+    }
+    var depPrimary = assetToPrimary[dep];
+    if (depPrimary != null) {
+      cleanedDeps.add(depPrimary);
+    } else {
+      log.info('Unable to find module for dependency: $dep '
+          'Are you missing a dependency on package:${dep.package}?');
+    }
+  }
+  return new Module(module.primarySource, module.sources, cleanedDeps);
+}
+
+/// Processes a [MetaModule] and adds to the [primaryToClean] cache.
+Future<Null> _cacheCleanedModules(
+    BuildStep buildStep,
+    MetaModule meta,
+    Map<AssetId, Module> primaryToClean,
+    Map<AssetId, AssetId> assetToPrimary) async {
+  _cacheAssetToPrimary(meta, assetToPrimary);
+  for (var module in meta.modules) {
+    primaryToClean[module.primarySource] = await _cleanModuleDeps(
+        buildStep, module, assetToPrimary,
+        isCourseStrategy: true);
+  }
+}
+
+enum Strategy { fine, course }
+
 /// Creates `.module` files for any `.dart` file that is the primary dart
 /// source of a [Module].
 class ModuleBuilder implements Builder {
-  const ModuleBuilder();
+  final bool _isCoarse;
+  const ModuleBuilder({bool isCourse: false}) : _isCoarse = isCourse;
+
+  static Strategy _getStrategy(BuilderOptions options) {
+    var config = (options.config['strategy'] ?? 'coarse') as String;
+    switch (config) {
+      case 'coarse':
+        return Strategy.course;
+      case 'fine':
+        return Strategy.fine;
+        break;
+      default:
+        throw 'Unexpected ModuleBuilder strategy: ${options.config['strategy']}';
+    }
+  }
+
+  factory ModuleBuilder.forOptions(BuilderOptions options) {
+    return new ModuleBuilder(
+        isCourse: _getStrategy(options) == Strategy.course);
+  }
 
   @override
   final buildExtensions = const {
@@ -24,15 +128,38 @@ class ModuleBuilder implements Builder {
 
   @override
   Future build(BuildStep buildStep) async {
-    if (!await buildStep.resolver.isLibrary(buildStep.inputId)) return;
+    Module outputModule;
+    var assetToPrimary = await buildStep.fetchResource(_assetToPrimary);
+    if (_isCoarse) {
+      var readMetas = await buildStep.fetchResource(_readMetas);
+      var primaryToClean = await buildStep.fetchResource(_primaryToClean);
+      var asset =
+          new AssetId(buildStep.inputId.package, 'lib/$metaModuleExtension');
+      // Even though the meta file may have already been processed call canRead
+      // to ensure it is properly marked as a dependency.
+      await buildStep.canRead(asset);
+      if (!readMetas.contains(asset)) {
+        var meta = new MetaModule.fromJson(
+            json.decode(await buildStep.readAsString(asset))
+                as Map<String, dynamic>);
+        await _cacheCleanedModules(
+            buildStep, meta, primaryToClean, assetToPrimary);
+        readMetas.add(asset);
+      }
+      outputModule = primaryToClean[buildStep.inputId];
+    } else {
+      if (!await buildStep.resolver.isLibrary(buildStep.inputId)) return;
 
-    var library = await buildStep.inputLibrary;
-    if (!isPrimary(library)) return;
+      var library = await buildStep.inputLibrary;
+      if (!isPrimary(library)) return;
 
-    var module = new Module.forLibrary(library);
-
+      var module = new Module.forLibrary(library);
+      outputModule = await _cleanModuleDeps(buildStep, module, assetToPrimary);
+    }
+    if (outputModule == null) return;
+    if (outputModule.primarySource != buildStep.inputId) return;
     await buildStep.writeAsString(
         buildStep.inputId.changeExtension(moduleExtension),
-        json.encode(module.toJson()));
+        json.encode(outputModule.toJson()));
   }
 }