+++ build_modules/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.2
+
+- Add new `strategy` option for `ModuleBuilder`. If `coarse` is provided a
+  minimum number of modules will be created. Otherwise, a module will be
+  created for each strongly connected component. Note that `coarse` is
+  the default strategy.
+
 # 0.2.1
 
 - Give a guaranteed reverse dependency order for